增加文本转语音模块Speaker、蜂鸣器Beep方法，并改运行方式为多线程

+++ Concentrating/KeyboardHelper.cpp
@@ -1,23 +1,40 @@
 #include "KeyboardHelper.h"
 #include <hook.h>
+#include <qthread.h>
 
 bool KeyboardHelper::_isLocked = false;
 
+QMutex KeyboardHelper::_mutex;
+
 bool KeyboardHelper::lock()
 {
-    _isLocked = true;
-    return InstallKeyboardHook();
+    if (!_isLocked) {
+        _isLocked = true;
+        _mutex.lock();
+        bool res = InstallKeyboardHook();
+        _mutex.unlock();
+
+        return res;
+    }
+    return true;
 }
 
 bool KeyboardHelper::unlock()
 {
-    _isLocked = false;
-    return UninstallKeyboardHook();
+    if (_isLocked) {
+        _isLocked = false;
+        _mutex.lock();
+        bool res = UninstallKeyboardHook();
+        _mutex.unlock();
+
+        return res;
+    }
+    return true;
 }
 
 void KeyboardHelper::setKeyState(int key, bool enable)
 {
-    ::SetKeyStatus(key, enable);
+    SetKeyStatus(key, enable);
 }
 
 bool KeyboardHelper::keyState(int key)
